requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiosignal==1.2.0
anyio==3.7.0
asttokens==2.2.1
async-timeout==3.0.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.68.2
filelock==3.12.2
flake8==6.0.0
flit_core==3.9.0
fonttools==4.41.1
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.7.0
ipykernel==6.25.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.0.3
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
motor==3.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
odmantic==0.9.2
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.10
pyflakes==3.0.1
Pygments==2.15.1
pymongo==4.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
rich==13.4.2
riotwatcher==3.2.5
scikit-image==0.21.0
scipy==1.10.1
setuptools==67.8.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.14.2
sympy==1.12
tifffile==2023.7.18
tomli==2.0.1
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.4.0
urllib3==2.0.3
uvicorn==0.22.0
wcwidth==0.2.6
Werkzeug==2.3.6
wheel==0.38.4
yarl==1.9.2
zipp==3.15.0
